Added dist.sh for compilation
The script compiles. Checks for a directory called lib in the working
directory. The actual SweetHttpd jar file should be copied/linked into
there to compile against.

If it is there, the script will produce a jar
that can be put into the server's plugin folder and loaded at runtime to
process php!